feat(RELEASE-1744): modify uniqueTag generation
This commit changes the uniqueTag logic to always prefer
the value from the org.opencontainers.image.created label.
When this label is not present, the old logic is used.

The reported uniqueTag for runs containing the label
will transform the value to unix time as this will be
the scanners expected format.

Signed-off-by: David Moreno García <damoreno@redhat.com>