fix(RELEASE-2055): remove filtered_snapshot from results
With large snapshots, users are hitting kubectl error
in update-cr-status. This is because one of the saved
results contained the full filtered snapshot, including
metadata (env vars, labels).

This result was saved in the filter task. Upon discussion
in the team, it seems we never intended to have this result
and it was there by mistake. So let's remove it.

The filter task only needs to save the advisory details
if all images are already released, but if not, no results
are needed.

BTW, this was supposed to be fixed in RELEASE-1972 but it turns
out that fix never really worked. We may want to remove that code
later.

Signed-off-by: Martin Malina <mmalina@redhat.com>