fix(ROK-1335): specify build type in update-package-collection
In push-rpm-to-koji pipeline, the task is run after
promote-koji-draft-build which promotes the build,
i.e. makes it draft=False.

Previously, the update-package-collection task ran ofter promotion
failed because of no draft build available.
This was improved in 07c0dde but we can still do better,
i.e. specify whether we want a draft build or not.
We want to list draft builds only after 'import'
and list normal builds after 'promote'.

Signed-off-by: Jiri Popelka <jpopelka@redhat.com>