fix(KFLUXSPRT-6360): per arch tarballs collected
This change fixes an issue where the push-artifacts-to-cdn
pipeline combined all arches for a single OS into a tarball
to deliver to cdn. Each os and arch combination should be a
separate archive.

This change also supports maintaining nested directory
structures and exact binary file names as the teams
deliver them in the source container.

Signed-off-by: Scott Wickersham <swickers@redhat.com>