fix: apply-mapping issue with helm charts
In RELEASE-1996 we removed the condition to check for helm charts.
The goal was to fetch both annotations and env vars and labels
for all images. But while we can fetch annotations for all images
(using skopeo inspect --raw), we cannot do normal skopeo inspect
on helm images, it will fail with:

unsupported image-specific operation on artifact with type
"application/vnd.cncf.helm.config.v1+json"

So let's skip this for helm charts.

Signed-off-by: Martin Malina <mmalina@redhat.com>