fix(KFLUXSPRT-6363): rpm manifest needs sync
There is a rare corner case when a rpm manifest object
gets created in Pyxis, but the ContainerImage object would not
have the reference to it.

The upload_rpm_data python script was updated to add a check
for this and do an empty patch request to Pyxis so that a sync
is triggered on Pyxis side.

This includes these utils PRs:

- https://github.com/konflux-ci/release-service-utils/pull/604
- https://github.com/konflux-ci/release-service-utils/pull/605
- https://github.com/konflux-ci/release-service-utils/pull/607
- https://github.com/konflux-ci/release-service-utils/pull/613
- https://github.com/konflux-ci/release-service-utils/pull/614
- https://github.com/konflux-ci/release-service-utils/pull/616

Signed-off-by: Martin Malina <mmalina@redhat.com>